Keying policy

+++ src/client/transmission_net_client.cpp
@@ -37,7 +37,7 @@ void ClientSocket::send(std::iostream &data) {
     if (!wait_connected()) {
         return;
     }
-    QByteArray qdata(inBase.str().data(), inBase.str().size());
+    QByteArray qdata(inBase.str().data(), static_cast<int>(inBase.str().size()));
     _socket->write(qdata);
     _socket->flush();
     emit sent();
@@ -67,15 +67,14 @@ void ClientSocket::receive() {
 }
 
 void ClientSocket::init() {
-    _socket->connectToHost(_address, _port);
+    _socket->connectToHost(_address, static_cast<quint16>(_port));
     if (wait_connected()) {
         _status = OK;
     }
 }
 
 void ClientSocket::_state_change(QAbstractSocket::SocketState state) {
-    if (_status != NEED_INIT
-        &&
+    if (_status != NEED_INIT &&
         (state == QAbstractSocket::SocketState::UnconnectedState
          || state == QAbstractSocket::SocketState::ClosingState)) {
         _status = NEED_INIT;
